Refactored BlogStorage to ES6 class to support DI
